+++ index.js
@@ -17,14 +17,6 @@ var parse = url.parse
 var Url = url.Url
 
 /**
- * Pattern for a simple path case.
- * See: https://github.com/joyent/node/pull/7878
- * @private
- */
-
-var SIMPLE_PATH_REGEXP = /^(\/\/?(?!\/)[^?#\s]*)(\?[^#\s]*)?$/
-
-/**
  * Module exports.
  * @public
  */
@@ -101,29 +93,48 @@ function originalurl (req) {
  */
 
 function fastparse (str) {
-  // Try fast path regexp
-  // See: https://github.com/joyent/node/pull/7878
-  var simplePath = typeof str === 'string' &&
-    str.charCodeAt(0) === 0x2f /* / */ &&
-    SIMPLE_PATH_REGEXP.exec(str)
-
-  // Construct simple URL
-  if (simplePath) {
-    var pathname = simplePath[1]
-    var search = simplePath[2] || null
-    var url = Url !== undefined
-      ? new Url()
-      : {}
-    url.path = str
-    url.href = str
-    url.pathname = pathname
-    url.search = search
-    url.query = search && search.substr(1)
-
-    return url
+  if (typeof str !== 'string' || str.charCodeAt(0) !== 0x2f /* / */) {
+    return parse(str)
   }
 
-  return parse(str)
+  var pathname = str
+  var query = null
+  var search = null
+
+  // This takes the regexp from https://github.com/joyent/node/pull/7878
+  // Which is /^(\/[^?#\s]*)(\?[^#\s]*)?$/
+  // And unrolls it into a for loop
+  for (var i = 1; i < str.length; i++) {
+    switch (str.charCodeAt(i)) {
+      case 0x3f: /* ?  */
+        if (search === null) {
+          pathname = str.substring(0, i)
+          query = str.substring(i + 1)
+          search = str.substring(i)
+        }
+        break
+      case 0x09: /* \t */
+      case 0x0a: /* \n */
+      case 0x0c: /* \f */
+      case 0x0d: /* \r */
+      case 0x20: /*    */
+      case 0x23: /* #  */
+      case 0xa0:
+      case 0xfeff:
+        return parse(str)
+    }
+  }
+
+  var url = Url !== undefined
+    ? new Url()
+    : {}
+  url.path = str
+  url.href = str
+  url.pathname = pathname
+  url.query = query
+  url.search = search
+
+  return url
 }
 
 /**